Jacobian removed from objfunc
As in create_model.py the jacobian matrix is not calculated, this
parameter was removed from objfunc and then into the scipy root
function, where calibration is carried out with the Levenberg-Marquardt
algorithm, we set the function to estimate jacobian numerically.
Furthermore, some parameters as ftol, xtol, gtol and eps was set with
specific values.