Area condition removed and distance criteria added
To take potential circle candidates we removed the area condition. In contrast to the above, a distance criterion was added, where we measure the distance between centroids of contiguous contours. Taking this into account, the distance between centroids of the concentric circles should be approximately zero. Then finally, we take the first three contours with the shortest distance which represent our three targets.