Two binarize methods and detection evaluation
Option to use a global or (local) adaptive thresholding as a second argument of the function. As a third argument, a threshold area is passed to evaluate the success of detection. If the sum of the areas of detected objects is lower to the threshold then the process fails, otherwise succeeds.