Name, inputs and outputs of known_var.py changed
The principal objective of known_var.py was to use the camera calibration parameters to estimate (with the target) the probe pose and save it with the cross-wire coordinates of images where correctly the target was detected. Both probe pose and cross-wire coordinates were the output of this code. Now only the probe pose is computed and therefore the output is only a list of the probe pose. This is why the name of this file is changed and its outputs and inputs (now cross-wire points are not read as input in the code).