Some changes due to calibration object created
With the calibration class recently added in calibration.py, calib.py structure is changed or adapted in order to use the calibration class, and now in this code, the calibration process is implemented more briefly. create_model.py is removed due to it is no longer necessary to create a calibration model to be loaded and used in calib.py, this is because calib.py create directly the calibration model. This change implies that model.pkl is no longer necessary and created.