getPose3D split into two functions
Function getPose3D is now splitted into two functions: centers3D which
compute 3D coordinates of the centers of each concentric circle tagged
as Xo (point in the origin), Xx (point in the x-direction) and Xy (point
in the y-direction). getPose which was adapted to the changes in
centers3D and return the rotation matrix and translation vector of the
target.