Variable with the target pose stored as .npy
Target pose was stored as a list with pickle, where each element of the
list was a 4x4 array with the pose of each frame. Now, the target pose
is stored with numpy.save as .npy file as a Nx4x4 array, where N is the
total number of frame.